requirements.txt
django>=4.2.*, <=4.3  # 4.2 is the Long-Term Service version.
psycopg2==2.*
requests>=2.31.*, <3.0
djangorestframework>=3.14.*, <4.0
pyjwt[crypto]>=2.8.0
pillow==10.*
